wip - removed further classes and extracted records
