+++ src/Server/src/AuthTokenGenerator/Abstractions/IIdentityService.cs
@@ -4,6 +4,7 @@ using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 using IdOps.Abstractions;
+using IdOps.Models;
 
 namespace IdOps.Abstractions
 {
