+++ src/Server/src/AuthTokenGenerator/Models/TokenRequestParameter.cs
@@ -1,0 +1,6 @@
+﻿namespace IdOps.Models;
+
+public record TokenRequestParameter(string Name)
+{
+    public string? Value { get; set; }
+}
