+++ src/Server/src/AuthTokenGenerator/Models/TokenRequestResult.cs
@@ -1,0 +1,8 @@
+﻿namespace IdOps.Models;
+
+public record RequestTokenResult(bool IsSuccess)
+{
+    public TokenModel? AccessToken { get; set; }
+
+    public string? ErrorMessage { get; set; }
+}
