+++ src/Server/src/GraphQL/Client/RequestTokenPayload.cs
@@ -1,4 +1,5 @@
 ﻿using IdOps.Abstractions;
+using IdOps.Models;
 
 namespace IdOps.GraphQL
 {
